SVN_MERGE
Merged revisions 1039806,1039816,1039915,1039919,1040235,1040285,1040301,1040468,1040669,1040878 via svnmerge from 
https://vkrause@svn.kde.org/home/kde/trunk/KDE/kdepim

................
  r1039806 | winterz | 2009-10-24 18:51:28 +0200 (Sat, 24 Oct 2009) | 19 lines
  
  Recorded merge of revisions 1039804 via svnmerge from 
  https://svn.kde.org/home/kde/branches/kdepim/enterprise4/kdepim
  SVN_SILENT:
  
  ................
    r1039804 | winterz | 2009-10-24 12:49:20 -0400 (Sat, 24 Oct 2009) | 11 lines
    
    Merged revisions 1039801 via svnmerge from 
    https://svn.kde.org/home/kde/branches/kdepim/enterprise/kdepim
    
    ........
      r1039801 | winterz | 2009-10-24 12:42:02 -0400 (Sat, 24 Oct 2009) | 4 lines
      
      don't check for forgotten attachments in invitation email
      
      MERGE: e4 (already in trunk)
    ........
  ................
................
  r1039816 | winterz | 2009-10-24 19:34:38 +0200 (Sat, 24 Oct 2009) | 3 lines
  
  compile warnings--
  MERGE: none
................
  r1039915 | djarvie | 2009-10-25 00:16:48 +0200 (Sun, 25 Oct 2009) | 1 line
  
  Improve menu
................
  r1039919 | djarvie | 2009-10-25 00:27:49 +0200 (Sun, 25 Oct 2009) | 2 lines
  
  Fixes for audio-only alarms
................
  r1040235 | winterz | 2009-10-25 22:28:54 +0100 (Sun, 25 Oct 2009) | 18 lines
  
  Recorded merge of revisions 1040231 via svnmerge from 
  https://svn.kde.org/home/kde/branches/kdepim/enterprise4/kdepim
  SVN_SILENT:
  
  ................
    r1040231 | winterz | 2009-10-25 17:26:50 -0400 (Sun, 25 Oct 2009) | 10 lines
    
    Merged revisions 1040223 via svnmerge from 
    https://svn.kde.org/home/kde/branches/kdepim/enterprise/kdepim
    
    ........
      r1040223 | winterz | 2009-10-25 17:20:17 -0400 (Sun, 25 Oct 2009) | 3 lines
      
      better context-aware messages when asking to send email for adding/changing/removing invitations.
      MERGE: e4 (already in trunk)
    ........
  ................
................
  r1040285 | djarvie | 2009-10-26 00:31:24 +0100 (Mon, 26 Oct 2009) | 2 lines
  
  Display file paths instead of URLs for local files
................
  r1040301 | djarvie | 2009-10-26 02:10:26 +0100 (Mon, 26 Oct 2009) | 2 lines
  
  Rename fileOrUrl() -> pathOrUrl()
................
  r1040468 | mlaurent | 2009-10-26 11:32:55 +0100 (Mon, 26 Oct 2009) | 2 lines
  
  Fix mem leak
................
  r1040669 | tokoe | 2009-10-26 17:44:50 +0100 (Mon, 26 Oct 2009) | 4 lines
  
  Do not use the time information in the CSV import/export dialog
  
  BUG: 207078
................
  r1040878 | djarvie | 2009-10-27 00:33:12 +0100 (Tue, 27 Oct 2009) | 2 lines
  
  Disable 'New Alarm from Template' action when no templates exist
................

svn path=/branches/work/akonadi-ports/kdepim/; revision=1041061

+++ kohelper.cpp
@@ -34,7 +34,7 @@ using namespace Akonadi;
 
 QColor KOHelper::getTextColor( const QColor &c )
 {
-  float luminance = ( c.red() * 0.299 ) + ( c.green() * 0.587 ) + ( c.blue() * 0.114 );
+  double luminance = ( c.red() * 0.299 ) + ( c.green() * 0.587 ) + ( c.blue() * 0.114 );
   return ( luminance > 128.0 ) ? QColor( 0, 0, 0 ) : QColor( 255, 255, 255 );
 }
 
@@ -56,6 +56,6 @@ QColor KOHelper::resourceColor( const Item &item ) {
 
 qint64 KOHelper::yearDiff( const QDate &start, const QDate &end )
 {
-  return start.daysTo( end ) / 365.25;
+  return static_cast<qint64>( start.daysTo( end ) / 365.25 );
 }
